Add ability to override dependency versions
This change extends rpc-differ to support the osa-differ argument
--version-mappings which was introduced in version 0.3.10. It allows a
version mapping to be specified for a project or role. For example, if
the project foo is pinned to the version 1.0.0 in RPC-OpenStack and the
project removes the tag and replaces it with v1.0.0, without this change
the command will fail. To address this, this change would allow the
following the be specified `--version-mapping 'foo;1.0.0:v1.0.0'`. A new
commit-ish can be proposed to override the one specified in
RPC-OpenStack to allow rpc-differ to work with versions of RPC-OpenStack
that no longer specify valid project/role versions.

Version updated to 0.3.10 in preparation for a new release.

JIRA: RE-2208